requirements.txt
certifi==2023.11.17
influxdb-client==1.38.0
inotify==0.2.10
nose==1.3.7
python-dateutil==2.8.2
python-dotenv==1.0.0
reactivex==4.0.4
schedule==1.2.1
six==1.16.0
typing_extensions==4.8.0
urllib3==2.1.0
